pyproject.toml
[project]
name = "zero-shot-realignment"
version = "0.1.0"
description = "Zero-shot transfer of steering-vector corrections across emergent-misalignment organisms"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "torch>=2.2.0",
    "transformers>=4.44.0",
    "datasets>=3.0.0",
    "accelerate>=1.0.0",
    "peft>=0.13.0",
    "bitsandbytes>=0.44.0",
    "trl>=0.12.0",
    "safetensors>=0.4.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0",
    "jsonlines>=4.0.0",
    "wandb>=0.18.0",
    "python-dotenv>=1.2.2",
    "nnsight>=0.6.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
    "ipython>=8.0.0",
]
eval = [
    "anthropic>=0.39.0",
    "google-generativeai>=0.8.0",
    "scipy>=1.14.0",
    "scikit-learn>=1.5.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "openai>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
